requirements.txt
# ============================================================================
# Intent-Recognition 依赖清单
# ============================================================================
# 安装: pip install -r requirements.txt
# 推荐: Python 3.11+
# ============================================================================

# ─── LangGraph 编排框架 ───
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0
langchain-core>=0.3.0
langchain-openai>=0.2.0          # 统一 OpenAI-compatible LLM 调用

# ─── Web 框架 ───
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1.0                    # FastAPI 模板引擎
python-multipart>=0.0.6          # FastAPI 文件上传支持

# ─── HTTP 客户端 ───
requests>=2.31.0                 # Dify API 调用 (同步)
aiohttp>=3.9.0                   # A2A JSON-RPC 调用 (异步)

# ─── 数据 & 配置 ───
pydantic>=2.0.0
pyyaml>=6.0                     # YAML 配置解析
python-dotenv>=1.0.0             # .env 文件加载
